requirements.txt
Django==4.2.1
djangorestframework==3.15.0
mysqlclient==2.2.7  
django-mysql==3.2.0
django-cors-headers==3.14.0
pika==1.3.1
psycopg2==2.9.10
cloudinary==1.42.2
django-cloudinary-storage==0.3.0
dj-database-url
python-decouple